Setup sound indicator
