--- js/components/AudioToggle/AudioToggle.module.scss
@@ -1,10 +1,0 @@
-.wrapper {
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    z-index: var(--z-layer-above);
-}
-
-.inner {
-    position: relative;
-}
