+++ js/modules/index.js
@@ -1,4 +1,3 @@
-export { default as AudioToggle } from "./AudioToggle";
 export { default as FadeInText } from "./FadeInText";
 export { default as FancyImage } from "./FancyImage";
 export { default as HomeExperience } from "./HomeExperience";
@@ -8,4 +7,5 @@ export { default as Loader } from "./Loader";
 export { default as PopQuote } from "./PopQuote";
 export { default as ScrambleText } from "./ScrambleText";
 export { default as Soundboard } from "./Soundboard";
+export { default as SoundIndicator } from "./SoundIndicator";
 export * as test from "./test";
